Refactor SubtitleBuilder.kt SubtitleBuilderTest.kt
* move SubtitleBuilderTest.kt to video package
* SubtitleBuilder.kt : reformat code
